hiberman: Set resume start timestamp only in the resume path
The resume start timestamp is currently set in setup_snapshot_device(),
this function is dedicated to setting up the snapshot device, setting
the timestamp is completely unrelated. Besides the function is also
called in the regular non-resume path. Set the timestamp after the
decision was made to resume, which is shortly before
setup_snapshot_device() is called.

BUG=none
TEST=resume time metrics are still in the same ballpark

Change-Id: I3c43e95f61005a681092a41ccca51dcffabab681
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4640450
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>