hiberman: Parallelize image loading and waiting for the system to be ready for resume
Loading the hibernate image takes several seconds, and so does waiting
for the system to be ready for resume. Currently these steps are
performed sequentially, so both contribute significantly to the total
resume time. It's not really necessary to wait for the system to be
ready for resume before loading the image. Instead launch a thread
that waits for the 'resume ready' signal before loading the image.
After the image is loaded join the thread (which usually will have
received the signal at that time).

BUG=b:288599039
TEST=hibermante hibernate -r
  => chrome://histograms/ResumeTime.Total is about 2s less

Change-Id: I5d6e3297fd1807bf835951a11fcee2b56a69288c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4676612
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>