patchpanel: ignore DNS redirection rule for Cellular multiplexing
This patch is a revert of crrev/c/4629328 (partial), crrev/c/4629329
(total) and crrev/c/4644888 (total) which incorrectly tries to swap in
DNS redirection rule requests the name of virtual network interfaces
used for VM or for ARC with the interface name of a shill Device.

BUG=b:273741099
TEST=FEATURES=test emerge-$BOARD patchpanel. Started patchpanel,
observed no error in logs. Observed that all DNS redirection rules with
a defined input ifname argument are for virtual network interfaces used
for VMs and ARC.

Change-Id: Ic2f51ce14c6f13e223e8317d8226915f12571c97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4644832
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>