cryptohome: Disable listing of legacy fp as a listed factor
Chrome doesn't really expect to see this factor listed as configured or
supported so we currently can't list it.

The implementation is rather unfortunate as we need to hard-code this as
a separate list directly in ListAuthFactors. However, we can't really
disable this at the driver level because that breaks the actual support
(which we expect to work and which we have tests for). So we kind of
need this half-measure to work around that.

This was "broken" by the changes a month ago or so that refactored all
the auth factor support into the new driver-based model. That shifted
both ListAuthFactors and the underlying authentication code over to
using the driver to detect support, whereas before they were somewhat
out of sync (the backend supported it but List ignored it). This change
recreates that distinction, although in a more explicit way.

BUG=chromium:1456863
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I2b41498264b2428037f4ed26ecd379316e2cd57d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4646853
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>