cryptohome: Handle overflow in recovery ID rotation
This treats overflow on recovery ID rotation as an error instead of
triggering undefined behavior.

While I'd say this is a strict improvement on the existing behavior, I
worry that this is not really the best way to handle this because once
you run out of increments rotate will just fail. On the other hand, that
requires you do rotate the ID 2**31 times and so not exactly all that
likely in practice.

Ideally the increment in the proto would've been a bignum like the seed
was, then we'd just increment this, but unfortunately since the protobuf
is persisted that's not an easy thing to fix. A cheaper solution might
be to instead promote it to an int64 which is somewhat backwards
compatible (new code can read the old protos) which in theory has the
same problem except that 2**63 increments is basically impossible.

BUG=b:287372375
TEST=re-ran fuzzer case

Change-Id: I6b063a9e6a80d076a76a1e32a9855316b7dc0551
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4648148
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>