cryptohome: fix a race condition on biometrics signal conversion.
Cryptohome currently skips signal conversion when the biometrics
enroll/auth session is not active. However, there is a tiny
time window when the enroll/auth session of biod has been
active, while cryptohome is still processing the session start
and yet to flip the state.

During that time window, the enroll/auth signals will be skipped
by cryptohome. This leads a race condition.

The race condition is discovered by the development of the tast
tests in crrev/c/4622160.

To fix this race condition, and not allowing the signal
conversion to happen unconditionally outside of active sessions,
we allow signal conversion when an biometrics enroll/scan session
is pending.

BUG=b:288966791
TEST=First set of fp tast tests pass and new unit tests pass.

Change-Id: Ideb69466e09843c7545437894ce056d1ba48663b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4646859
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: Zi Lin <lziest@google.com>
Tested-by: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>