rmad: Read segmentation values in UpdateDeviceInfo state
Get the feature level of the device and pass it to Chrome. Currently we
don't set GSC values when the feature level is UNKNOWN. Also provide a
way to inject feature values so we can test all kinds of scenarios on a
single device.

BUG=b:286502404
TEST=FEATURES=test emerge-dedede rmad

Change-Id: I7e22fa74eb05f48861d91faa31afde4322be4543
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4647077
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Jeff Lin <jeffulin@google.com>