vm_tools/sommelier: separate namespace for Wayland read pipes
talex5@ reports:

"When sending a pipe, the guest has to know what ID the host will
assign to it. The guest guesses, by adding 2 to the last ID it knows
about, and the host checks its guess. If it was wrong
(e.g. because the host created a new ID at the same time) then
you get a protocol error."

A solution is a separate namespace for Wayland read pipes.

BUG=b:259268477
TEST= copy+paste still works

Change-Id: I996b56a38b4794a252829cf48d85e1ead58dee8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4653983
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Auto-Submit: Gurchetan Singh <gurchetansingh@chromium.org>