patchpanel: introduce absence of default shill Devices (1/2)
When switching the tracking of Cellular Devices to take into account the
primary multiplexed interface and to ignore Cellular Devices without a
primary multiplexed interface, patchpanel handlers for default logical
and physical Devices events must not be triggered because the network
interface is not known.

To prepare for this transformation, this patch makes ShillClient
properly notify when the shill default logical Device or default
physical Device are not defined, and updates all use sites to handle
this scenario. This patch does not change how ShillClient tracks
internally the default logical and physical Device.

BUG=b:273741099
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: Ib494608bf9938da4cfbb9721bd76439563eacb7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4656481
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>