libec: Change error/warning to show cmd as hex
This allows you to easily search for the hex number
in the EC's ec_commands.h file, since they are all
defined in hex.

Also, the rest of this source file prints the command
as hex. This is the only place where we don't print
as hex.

BUG=b:288463914, b:287268819
TEST=# With SW flash protect already disabled, but HW WP enabled
     ectool --name=cros_fp flashprotect enable

Change-Id: I6ddd1d56ed0ffb9c740b3d422637e8bd4fb36f30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4659696
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>