cryptohome: Get rid of auth session status
The currently auth session status basically does two separate things: it
tracks if the session has been authorized for decrypt, and it tracks if
the session has been timed out.

For the decrypt check, we can instead switch to checking the authorized
intents directly. For the timeout we can replace this with a separate
flag which is indepenedent of the authorized intents.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ia3e69e9479e0a690eb87d5f20cbb94f7edd95059
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4659861
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>