oobe_config: Add metrics lib methods to FileHandler
Create methods required for the Rollback metrics library. Usage is
added in follow-up CL.

Create method to be able to create a file atomically containing the
provided data.

Create methods to be able to lock the file using flock. By design,
we will prefer to lose data than blocking Rollback. Therefore, if
the lock is busy, we do not wait for it to be released.

Create methods necessary for testing.

See details about the file sync implementation decisions in
go/cros-rollback-metrics-data-file-sync.

Create separate BUILD.gn for filesystem for easier readability.

BUG=b:261850979
TEST=1)FEATURES=test USE="asan fuzzer" emerge-amd64-generic oobe_config
     2)tast -verbose run DUT rollback.EnterpriseRollbackInPlace
     3)tast -verbose run DUT rollback.OobeConfigSaveShutdown

Change-Id: I79c505d75cb65c70ac58b731101aa84c9d3a62ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4350020
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>