arc-setup: Ensure esdfs mount points are unmounted in login screen
esdfs mount points (/run/arc/sdcard/*/emulated) are unmounted in the
post-stop stanza of arc-sdcard, which is currently stopped from the
post-stop stanza of arc-lifetime.

With this setup, when arc-liftime is stopped on `stopped ui`, it is not
guaranteed that arc-sdcard is already stopped at the timing where the
login screen is displayed, and thus can leak the esdfs mount points in
the login screen.

To avoid such a leak, this CL 1) sets arc-sdcard's stop condition to
`stop on stop-arc-instance or stopping ui` and 2) sets arc-lifetime's
stop condition to `stop on (stop-arc-instance and stopped arc-sdcard)
or stopped ui`.

arc-sdcard's stop condition gurantees that the esdfs mount points are
unmounted when the login screen is displayed, and arc-lifetime's stop
condition guarantees that the esdfs mount points are unmounted when
"arc-setup --mode=stop" is executed (otherwise the command fails to
unmount sdcard_mount_directory).

Note that setting arc-sdcard's stop condition to `stop on stopping
arc-lifetime and stopping ui` does not guarantee the above conditions;
if stop-arc-instance is emitted while stopping ui, arc-lifetime's
post-stop stanza can be executed before that of arc-sdcard.

BUG=b:288954735
TEST=tast run DUT '(!informational && "group:mainline" && "name:arc.*")'
TEST=Insert a 10-second sleep at the beginning of UnmountSdcard(), run
TEST=the following commands, and confirm that arc-lifetime's post-stop
TEST=stanza is always executed after that of arc-sdcard:
TEST=stop ui
TEST=initctl emit stop-arc-instance
TEST=stop ui -> (while stopping ui) initctl emit stop-arc-instance

Change-Id: I82dd0b081b6de67c3ffdc99c4a5c5cd3f6b14010
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4664283
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Yury Khmel <khmel@google.com>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>