oobe_config: Create EnterpriseRollbackMetricsHandler
Create shared library EnterpriseRollbackMetricsHandler that allows
to track the events triggered during Enterprise Rollback so they
can successfully be reported after powerwash.

Ensure that the rollback metrics file is only created if metrics
are enabled. Events are only included in the file if it already
exists.

The library can be called from different processes. Ensure the file
is created containing the right rollback metadata. Use flock when
adding new events to the file and when reporting them. By design,
we prefer to lose metrics than blocking Rollback. If the lock is
busy, we do not wait for it to be released.

Structured metrics reporting has not been implemented yet.

Create separate BUILD.gn for metrics for easier readability.

BUG=b:261850979
TEST=1) FEATURES=test USE="asan fuzzer" emerge-$BOARD oobe_config
     2) tast -verbose run DUT rollback.EnterpriseRollbackInPlace
     3) tast -verbose run DUT rollback.OobeConfigSaveShutdown
     2) Local test. Use shared library in update engine

Change-Id: I1199158d9656c8e7042d78f83510d281bdbe4a49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4661000
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>