screen-capture-utils: Screen rotation support
Frame buffers of devices like kukui are in portrait layout.
When those devices are used in landscape mode, VNC will
display the screen conetnts rotated 90 degrees, which is
quite confusing and difficult to use.

This patch adds --rotate flag to kmsvnc and screenshot
to solve this problem. Per the investigation documented
in the bug, auto detection of the screen orientation
appeared to be non-trivial, thus the CL does not
support it.

Rotation by 180 and 270 degrees are intentionally not
supported in the spirit of YAGNI.

BUG=b:177967871
TEST=Manually test with kukui and xtigervncviewer; unit tests

Change-Id: I5ca49f7ef1c1c055753d399693f01e60a22ed09f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4666587
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Satoru Takabayashi <satorux@chromium.org>