oobe_config: Add ReportEventNow to metrics library
Create method that allows to report the event immediately instead
of adding it to the rollback metrics data file. This method must
only be called for rollback events that happen after powerwash.

ReportEventNow requires access to the rollback metrics data file
to retrieve the metadata and use it for reporting the structured
metrics. To minimize the lose of data, this method also attempts to
report the old metrics tracked in the file.

Structured metrics reporting has not been implemented yet.

BUG=b:261850979
TEST=1) FEATURES=test emerge-amd64-generic oobe_config
     2) tast -verbose run DUT rollback.EnterpriseRollbackInPlace
     3) tast -verbose run DUT rollback.OobeConfigSaveShutdown

Change-Id: I80356b02d8e6a28498c62e3dc6d56598aab2c115
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4664800
Reviewed-by: Igor <igorcov@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>