cryptohome: Don't recalculate PW root hash by default in Sync
Currently we `sync` before calling all PW operation in LECredManager,
which does the following:
    1. Check whether root hash produced by leaf hashes is equal to
       root_hash retrieved from latest PW operation's response.(this
       takes latency of a lot of hashes)
    2. If 1. is not successful, try loading leaf hashes from disk again
       and repeat 1.
    3. If 2. is still not successful, do sync replay.

This CL made change to step 1 so now it doesn't recalculate root hash cache (flow of step 2, 3 remain unchanged). This makes sense since we
always update the inner hash cache (up to the root hash) after we
store/remove a label. If the cache is out of sync then steps 2 or 3 can
still recover the cache.

The main advantage is that we can save the latency for calculating the
root hash for most of the case.

BUG=b:289489750
TEST=cryptohome unittests
TEST=tast test dut hwsec.PINWeaver cryptohome.AddRemovePIN
TEST=performance gain: TBD

Change-Id: I374d086d8928fdcd612e4d6689bb5a586bdf6c74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4666592
Tested-by: Sam Chen <chensa@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Howard Yang <hcyang@google.com>
Auto-Submit: Sam Chen <chensa@google.com>
Commit-Queue: Sam Chen <chensa@google.com>