secanomalyd: adds android-data/data as known mount
/home/root/<hash>/android-data/data shows up as a writable and
executable mount for less than a second on container-based ARC++
systems. This CL adds this mount to the list of known mounts, to prevent
additional crash reports being sent for this mount. The mount is still
reported if it shows up in two consecutive scan intervals.

BUG=b:285028181
TEST=The mount is ignored when only present during a single scan.

Change-Id: I44661b21e72ad562b55f065198146fdf593c7f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4670170
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>