arc-setup: Stop arc-sdcard on stopping ui separately from arc-lifetime
Currently, arc-sdcard is stopped on the post-stop stanza of arc-lifetime
to ensure that the esdfs mount points (/run/arc/sdcard/*/emulated) are
unmounted before sdcard_mount_directory (/run/arc/sdcard) is unmounted.

However, arc-lifetime's stop condition contains `stop on stopped ui`,
which can be executed after the login screen is shown, and thus can leak
the esdfs mount points in the login screen.

To avoid such a leak, this CL makes arc-sdcard stop on
`stop-arc-instance or stopping ui`, which ensures that the esdfs mount
points are unmounted when the login screen is shown.

Also, to keep ensuring that the esdfs mount points are unmounted before
sdcard_mount_directory, this CL lets the esdfs mount points be unmounted
in `arc-setup --mode=stop` which is run from arc-lifetime's post-stop.
The mount points are unmounted via UnmountIfExists(), which can safely
be attempted multiple times.

This is an alternative to CL:4671487, which tried to solve the issue by
representing the dependency by tweaking stop conditions in a wrong way.
Calling UnmountIfExists() multiple times is redundant but robust against
all possible sequences of stop-arc-instance, stopipng ui and stopped ui.

BUG=b:288954735
TEST=arc.StartStop, arc.ManagedPlayStoreIcon
TEST=tast run '(!informational && "group:mainline" && "name:arc.*")'

Change-Id: I1e518a22bdbc00c3546084ec7188c569f299d446
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4671069
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Reviewed-by: Vaibhav Raheja <vraheja@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>