dlp: Start watchdog on fanotify thread
If the main thread is already non-responsive,
than upon receiving a fanotify open permission request on
fanotify thread, it won't be able to post a processing
task to the main thread (that would start a watchdog).

Instead, the watchdog should be created in fanotify
thread and ensure that it's responded on time (1 sec)
(in FanotifyWatcher::OnRequestProcessed()).

Otherwise, the watchdog will crash DLP daemon process
causing it to restart.

BUG=b:290187330
TEST=FEATURES=test emerge-{$BOARD} chromeos-base/dlp
TEST=tast run <DUT> dlp.DataLeakPreventionRulesListFiles*
TEST=Artificially stop responding in the daemon and check that it's restarted.

Change-Id: I01c19a96bd11a71601107af3b648a2b52c5de357
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4669449
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Reviewed-by: Aida Zolic <aidazolic@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>