shill: Make parsing functions a members of an endpoint
Currently ParseIEs and ParseVendorIEs functions of WiFiEndpoint are
static functions with many output arguments.  With addition to support
for transitional mode OWE AP there will be more output parameters needed
and this becomes "ugly".  So let's make the functions a member functions
and just store the values in internal data members (which is what
happens anyway because we are passing pointers to these members as
arguments).

BUG=b:289025277
TEST=USE="asan fuzzer" FEATURES=test emerge-$BOARD shill

Change-Id: I61fdd27f04d0ea5d3f54bafd1c3bb9d878251b5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4670346
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>