imageloader: Stop stanza update
When imageloader-shutdown is starting, no new instances of imageloader
service should running and be stopped.
This handles the case of stopping.

This is to stop all instances of imageloader runs during shutdown.

There should be no new instances of imageloader started, which will be
done with a dependency on a stanza that is running during the lifetime
of when pre-shutdown starts.
Previously guarding when `ui` job is down, was too limiting.

BUG=none
TEST=CQ

Change-Id: I1d6f906a9fe755fe35c4dd7b345b55fd9be49576
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4673310
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>