patchpanel: check if the default logical or physical Devices are active
This patch ensures that if a shill Device is not active, it cannot be
advertised as the default logical or physical shill Device inside
patchpanel.

Once Cellular multiplexing is enabled, this situation can happen if
Cellular is enabled but not connected (therefore the primary multiplexed
interface is undefined) and no other network is connected.

BUG=b:273741099
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: Ie63edb52de3d8524c945407e7d7662b08b54ba54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4699523
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>