libsegmentation: Align enums with feature-management
Unify the conversion of "feature scope" and "feature level" to protobuf
values in feature-management, and sync the "feature usage" enum.

We guarantee with a unit test that scopes defines from the
feature_management proto and the internal libsegmentation proto are
identical, we can compare both freely.

Regenerate the testing protobuf message with the updated definition.

BUG=b:273883217
TEST=unit test

Cq-Depend: chromium:4619555
Change-Id: Iaa9a629db7fa89167434f247b0b1a7b6db6622fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4661929
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>