cryptohome: Get rid of set_features in session manager
There's not really any reason to asynchronously pass in
AsyncInitFeatures to a class; the whole point of the AsyncInit class is
to do the "it gets set later" logic for you, so this is just making the
init...double async.

The only real downside is that this forces code which is never planning
to initialize it to pass in a real async init features. But given that
no tests were taking advantage of this (every test that constructs
AuthSessionManager makes an AsyncInitFeatures and passes it in via
set_features) it's not actually helping.

BUG=b:290679363
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic3862464a5fb03e65ce16bd93e8585c37e407141
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4669404
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>