vm_tools: vm_protos: Sync VmType enum in vm_host.proto.
The VmType enum in vm_tools/proto/vm_host.proto should be the same as
the one in system_api/dbus/vm_concierge/concierge_service.proto. The
enum in vm_host.proto is used by the Vmlog Forwarder and then forwarded
to Anomaly Detector.

If the enum is not synchronized, the Anomaly Detector won't be able to
identify the specific VM from which the logs originated.

BUG=b:286479507
TEST=emerge vm_protos vm_host_tools
TEST=See this log when starting Borealis. Previously, type was empty.
TEST=Received VmStartingUpSignal for owner_id/borealis, cid 41, type BOREALIS

Change-Id: Idd8732342257feb5aed05bcae68def4ae67b24d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4675019
Reviewed-by: Phillip Pearson <philpearson@google.com>
Commit-Queue: Norman Bintang <normanbt@chromium.org>
Tested-by: Norman Bintang <normanbt@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>