vm_tools: Move *_dev.conf documentation to markdown.
arcvm_dev.conf comment was growing to be more than 4kB in size, and I
don't want to use more than 1 page for this configuration file. We don't
need the file to be installed on the device, we can read it in the
source tree or the web site.

BUG=b:289939934
TEST=read it.

Change-Id: I7c494cdac53d8311d4a95ffd2f10d2c735935829
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4672294
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Tested-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>