shill: cellular: keep track of the last APN type connected as default PDN
The logic to connect a DUN APN as DEFAULT packet data network will
create connected Bearers with APN type DUN which need to be considered
as if they were of DEFAULT type during the whole cellular connection
sequence.

In order to allow that, we will keep track of which is the last APN
type used to connect the |default_pdn_| so that we know if it was a
DUN APN or a DEFAULT one, and be able to lookup the correct Bearer
accordingly.

BUG=b:283396208
TEST=cros_workon_make --board=${BOARD} --test chromeos-base/shill

Change-Id: Ie2bbee9da94b68b6c5b68737043d3fe2321812a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4677336
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>