cryptohome: Lots of DCHECK -> CHECK
This changes a lot of DCHECKs to CHECKs. In order to keep this change
"safe" it just focuses on really obvious checks where the "D" is
providing no benefit, namely null checks (and optional::has_value
checks). In those cases the subsequent code would just crash anyway (if
we're lucky; it's actually undefined behavior...) and so it's better to
just CHECK anyway.

This mostly doesn't do anything with the other DCHECKs that were
checking more complex conditions. They arguably should be cleaned up as
well (either to CHECKs or downgraded to something that's testable) but
this change is more low hanging fruit.

BUG=b:290817945
TEST=FEATURES=test emerge-$BOARD cryptohome cryptohome-dev-utils

Change-Id: I73f55c930d746df4c1137fab68535187a94de5de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4678862
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>