secanomalyd: keep mounts/processes lists in-memory
Adds the list of all mounts and all processes as daemon attributes. This
is needed to facilitate a detection and reporting model where we scan
for both process and mount anomalies before sending a report without
worrying about changes in the state of the system by the time we are
ready to send a report.

Changes the detection and reporting model such that the reporting is
triggered based on the presence of anomalies after both mount and
process scanning happens and the information in the reports is taken
from when the anomaly was detected.

Removes the concept of filters when it comes to reading mounts, and adds
a method for filtering out private mounts, instead of filtering as we
read.

BUG=b:290827485
TEST=Reports include the list of processes and mounts at the time of
detection.
TEST=All unit tests pass.

Change-Id: I4ee910fa935e752e10475ba89f17576b7831431a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4679907
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>