shill: cellular: no src address for entitlement checks
This patch skips the configuration of the src IP address of libbrillo
http / curl for Cellular tethering entitlement checks and replaces it
with configuration the network interface to bind to. This allows to
defer the selection of the IP family to curl, based on available DNS
servers.

BUG=b:285242955
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ib55162a4d0c141063afb987bc555c1f077d8e939
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4680641
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>