power: Deflate PeripheralBatteryWatcherTest.SpammyUdevEvents
PeripheralBatteryWatcherTest.SpammyUdevEvents was triggering udev events
without waiting for them to complete. If the udev event was delivered
before the test completed, it would cause a CHECK() failure since the
event wasn't being waited for. Most of the time the test ran quickly
enough for this to happen, but sometimes we would lose the race.

Wait for the event explicitly to fix the bug, and also reduce the number
of iterations to reduce the impact on test run time.

BUG=b:289969436
TEST=cros_sdk FEATURES=test emerge-nissa power_manager

Change-Id: I27b85f54d5cfe20f15b2ff465021016317b79c2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4678561
Commit-Queue: Nancy Zhao <zhaon@google.com>
Auto-Submit: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Tested-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Nancy Zhao <zhaon@google.com>