mount-passthrough: Add support for casefold lookup
In Android R, case insensitive lookup support on the My files volume is
checked by CTS.

On ARCVM, the feature is supported by virtio-fs, but on R container, we
need to support it in another way.

This CL adds a new command line option for mount-passthrough to enable
case insensitive lookup. The feature will be enabled by another CL only
on My files on R container.

Note that the feature should not impose much performance cost as long as
the guest accesses existing files with their exact paths, which is
expected to be the majority of file accesses to the My files volume.

BUG=b:286238422
BUG=b:291063439
TEST=See CL:4680871
TEST=arc.MyFiles(.vm), arc.RemovableMedia(.vm)
TEST=FEATURES=test emerge-${BOARD} mount-passthrough
TEST=Manually confirm that the feature is not enabled by default.

Disallow-Recycled-Builds: test-failures
Change-Id: I6fe7728aa10426628ea45b8bc6262e99829a8ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4680870
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>