cryptohome: Move auth session timeout logic to the manager
Move all of the support for timing out auth sessions out of the session
objects themselves and into the session manager. From a logical
perspective this is reasonable (the timeout is already linked to
lifetime, which the manager is responsible for) and it simplifies the
state in the session itself because we no longer need to maintain a
separate "no longer authenticated" state distinct from "not yet
authenticated".

BUG=b:290679363
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Icf37ca40aa204c2905f2ef39d60ead263ec6c81d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4687231
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>