common-mk: Avoid modifying the original sysroot on setup
To prepare for an upcoming patch to support the unprivileged strategy,
this patch improves platform2_test.py to not modify the original sysroot
before entering it. Currently it is possible to modify the sysroot
because the process gets the root privileges by sudo, but it won't be
possible when not using sudo.

BUG=b:262458823
TEST=./platform2_test.py --sysroot=/build/arm64-generic bash

Change-Id: Ibe2e1e5fb82f0fa12643dbb7a5175fb13632dd6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4680609
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>