vm_tools: concierge: Factor out boot clean-up logic to ArcVm from service
Currently, the request handler of ArcVmCompleteBoot has the knowledge of
the necessary clean up of an ArcVM instance and modifies the VM state
based on that. This change makes ArcVmCompleteBoot delegate the VM clean
up to ArcVM class. Note this is also a preparation for adding boot
metrics to ArcVM in following changes.

BUG=b:199825292,b:286164974
TEST=`cros_run_unit_tests --packages vm_host_tools --board=brya-arc-t`

Change-Id: Ibc9c89caffff96e59c61ab26ba91a27f45ac5b2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4684568
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Masami Hiramatsu <mhiramat@google.com>
Commit-Queue: Takaya Saeki <takayas@chromium.org>
Tested-by: Takaya Saeki <takayas@chromium.org>