crash: Refactor collectors to share MetricsLibrary object
Add MetricsLibrary parameter to collector constructors. CrashReporter
will create a single MetricsLibrary object that will be shared with all
other collectors.

This is necessary for future work that will instantiate MetricsLibrary
with a nonblocking MetricsWriter to mitigate the risk of CrashReporter
deadlocking when calling MetricsLibrary::Send*ToUMA methods in the event
that the crashed process is holding the metrics file lock.

BUG=b:270732207
TEST=cros_run_unit_tests --board=${BOARD} --packages
chromeos-base/crash-reporter

Change-Id: I662a7cfa0aa00f104e7e081710a517545b54e96e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4684507
Tested-by: Kendrake Tsui <kendraketsui@google.com>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>