shill: Add flag is_dbus_call to Device::Scan()
There is no flag about the caller in the parameters of Device::scan(),
so it doesn't know if it is called by Manager::RequestScan().

Add flag is_dbus_call to Device::Scan(), which is true if it is called
from D-Bus, for example, Manager::RequestScan().

BUG=b:290907571
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I3e7ca4075434d322d5937ec88148b15d437d8976
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4685428
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kaidong Wang <kaidong@chromium.org>
Tested-by: Kaidong Wang <kaidong@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>