shill: metrics: migrate NetworkServiceError
This patch migrates from the Metrics class to the Service class the
function that converts Service::ConnectFailure enum values to
NetworkServiceError enum values.

Because Service forward declares Metrics and because NetworkServiceError
is a nested enum, NetworkServiceError is migrated temporarily to
MetricsEnums.

This removes one direct dependency of Metrics on Service.

BUG=b:268579868
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I9a7f332fc19b7d5584416e1ffea2edd08d693b10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4685038
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>