shill: metrics: migrate UserInitiatedConnectionFailureReason
This patch migrates from the Metrics class to the Service class the
function that converts Service::ConnectFailure enum values to
UserInitiatedConnectionFailureReason enum values.

Because Service forward declares Metrics and because the metric enum
is a nested enum, it is migrated temporarily to MetricsEnums.

This removes one direct dependency of Metrics on Service.

BUG=b:268579868
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I11a2c19cea15e7fe6b15dccff2f1345c39391471
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4685039
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>